switched query and train img, draw camera pose estimate, best 80
captured a new image with the camera a little farther from the bag.
Best matches 80 worked best this time.

in "ORBmatches inverted", it shows the border of the captured image
overlayed on top of the original shoppingbag image. We also draw
the center of the overlayed border, which estimates the camera pose.

From the captured image, we can see that the camera is behind the
bottom-right hand corner of the bag (on the left)
If we look at the drawn estimate, we can see the same thing. The
camera pose estimate is below the bottom-right of the bag (on the right)